requirements.txt
asgiref==3.8.1
asyncio==3.4.3
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
click==8.1.7
dnspython==1.16.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
patsy==0.5.6
prophet==1.1.5
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
scipy==1.14.0
six==1.16.0
statsmodels==0.14.2
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.2
scikit-learn==1.5.2
redis==5.2.0